rename enchanted to enhancement
