--- src/app/simu/image_enchanted/page.tsx
@@ -1,13 +1,0 @@
-"use client";
-import BaseLayout from "@/components/layout/BaseLayout";
-import GrayscaleTransformSection from "@/components/image_enchanted/grayscaleTransform";
-import SharpenSmoothTransformSection from "@/components/image_enchanted/sharpenSmoothTransform";
-export default function ImageEnhancementPage() {
-
-  return ( 
-    <BaseLayout>
-      <GrayscaleTransformSection />
-      <SharpenSmoothTransformSection />
-    </BaseLayout>
-  ); 
-}
