+++ src/app/simu/image_enhancement/convolution/page.tsx
