+++ src/assets/templates/image_processing.html
@@ -102,7 +102,7 @@
     </main>
 
 
-    <script src="/page/js/image_processing/script-enchanted"></script>
+    <script src="/page/js/image_processing/script-enhancement"></script>
     <script src="/page/js/image_processing/script-noise"></script>
     <script src="/page/js/image_processing/script-rgb"></script>
 </body>
