+++ src/components/image_enhancement/data/types.ts
