+++ src/components/image_enhancement/logic/gray_scale.ts
