+++ src/components/image_enhancement/logic/noise.ts
