+++ src/components/image_enhancement/logic/process_image_convolution.ts
