+++ src/components/image_enhancement/nodes/GrayScaleNode.tsx
