+++ src/components/image_enhancement/nodes/ImageOriginalNode.tsx
