+++ src/components/image_enhancement/states/types.ts
