+++ src/components/image_enhancement/types.ts
