+++ src/utils/translations/imageenchanted/th.json
@@ -1,5 +1,5 @@
 {
-  "whatisimageenchanted": "Image Enchanted คืออะไร?",
-  "mean": "Image enchanted คือ การดำเนินการทางคณิตศาสตร์ที่ใช้เพื่อจัดการภาพโดยการนำเคอร์เนล (kernel) ทางคณิตศาสตร์ไปใช้กับแต่ละพิกเซลของภาพ เป็นการดำเนินการพื้นฐานในการประมวลผลภาพและ computer vision",
-  "mean_sub": "การ enchanted ของภาพคือกระบวนการดูภาพเป็นเมทริกซ์ของพิกเซลที่เรียงกันเป็นแถว ตัวอย่างเช่น เราอาจมี 1 สำหรับสีดำและ 0 สำหรับสีขาว จากนั้นเราจะใช้สิ่งที่เรียกว่าตัวกรอง (filter, kernel) เพื่อ enchanted ภาพ และผลลัพธ์จะเป็นเมทริกซ์สีแดง"
+  "whatisimageenchanted": "Image Enhancement คืออะไร?",
+  "mean": "Image enhancement คือ การดำเนินการทางคณิตศาสตร์ที่ใช้เพื่อจัดการภาพโดยการนำเคอร์เนล (kernel) ทางคณิตศาสตร์ไปใช้กับแต่ละพิกเซลของภาพ เป็นการดำเนินการพื้นฐานในการประมวลผลภาพและ computer vision",
+  "mean_sub": "การ enhancement ของภาพคือกระบวนการดูภาพเป็นเมทริกซ์ของพิกเซลที่เรียงกันเป็นแถว ตัวอย่างเช่น เราอาจมี 1 สำหรับสีดำและ 0 สำหรับสีขาว จากนั้นเราจะใช้สิ่งที่เรียกว่าตัวกรอง (filter, kernel) เพื่อ enhancement ภาพ และผลลัพธ์จะเป็นเมทริกซ์สีแดง"
 }
